requirements.txt
docling
htmldocx
python-docx
torch
googletrans